Move stdin tests to own file
